Update information board

+++ Assets/Scripts/Game/Models/ITower.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public interface ITower {
 	float damage { get; set; }
-	float attackSpeed { get; set; }
+	int attackSpeed { get; set; }
 	float critRate { get; set; }
 	float critFactor { get; set; }
 	float attackRange { get; set; }
